Prefactor workflowRunner for generating text zip
Moved all text/alto building to separate class
